Babel and some conf changes
The calculator variable is now exported as global and could be used from
the index.html/calc.html page directly.

As a side change a minor improvement was performed in the
webpack.config.js for the build and for the formatting.

Babel is now working
In dev server css can't be loaded

Included css file to be used by the dev server
Calculator is working properli with fractions

Added some test with karma runner using jasmine

Updated Babel

The calculator variable is now exported as global and could be used from
the index.html/calc.html page directly.

As a side change a minor improvement was performed in the
webpack.config.js for the build and for the formatting.